+++ tests/com.b2international.snowowl.test.commons/src/com/b2international/snowowl/test/commons/Resources.java
@@ -38,7 +38,7 @@ public class Resources {
 
 		public static final String SNOMED_RESOURCES = createResourcePath(RESOURCES, "snomed");
 		public static final String MINI_RF2_INT_20200131 = createResourcePath(SNOMED_RESOURCES, "SnomedCT_InternationalRF2_PRODUCTION_20200131T120000Z_minified.zip");
-		public static final String MINI_RF2_INT_20210131 = createResourcePath(SNOMED_RESOURCES, "SnomedCT_InternationalRF2_PRODUCTION_20210131T120000Z_minified.zip");
+		public static final String MINI_RF2_INT_20210131 = createResourcePath(SNOMED_RESOURCES, "SnomedCT_InternationalRF2_PRODUCTION_20210131T120000Z-minified.zip");
 		public static final String MINI_RF2_INT_20210731 = createResourcePath(SNOMED_RESOURCES, "SnomedCT_InternationalRF2_PRODUCTION_20210731T120000Z-minified.zip");
 		public static final String MINI_RF2_EXT = createResourcePath(SNOMED_RESOURCES, "SnomedCT_RF2Release_INT_20160501_B2i_cd_refsets.zip");
 		public static final String MINI_RF2_COMPLEX_BLOCK_MAP = createResourcePath(SNOMED_RESOURCES, "SnomedCT_RF2Release_GB1000000_20200204_complex_block_map.zip");
